Decoupled search/autocomplete componenets
